Controller - Movement
Continuing to work on making the movement system feel responsive
Tightened jumping, it now has much more forgiveness around when you
asked to jump verses when it will actually jump (it felt awful before,
even if it was the correct behaviour).

MoveXZ360
I am about to do some fairly comprehensive changes to this, so this is
my restore point, if you will. Also made this move support either having
air vectoring or being on the ground as necessary conditions for
movement.
I intend on making this function work such that the controller is turned
towards the desired direction of movement, but otherwise moves forward
when any input is received.