Capstone 1 Polish 2
I have added new fireworks, and clouds to the title scene.
I read up on how scene loading works, I will be able to make this work,
via loading the scene until it is ready and fading the music over the
loading time, or some amount of time, say an 8 second fade to black, or
to white, depending on time of day, then brighten on the Range.

So lots of cloud textures
A new environment xcf
Some more particles
Flare core
The fireworks are traily ones, they go up leaving a trail then when the
trail disappears they create a burst of flares that leave trails as well
that fade with the bursts, all quite pleasant.

Next I will create the transition system so I can start shunting the
player around scenes, get the Range up and running, and hopefully build
a second character, after I finish Nero, for Monday.