World
I made a new world. This one is cooler, it is similar but better fleshed
out, more useful in other contexts, as well this one will be getting a
texture over the next little while, I want it ultimately to look like a
landscape, so I need to put the time and effort in to make it look like
a landscape, trees, etc, the Originator is not a boring person, just a
dictator.