The Environment
I have made a more convincing cloud system, as well I have improved the
EnvironmentController by giving it EnvironmentWeatherSystems, which
contain weather state data, and the logic with which to execute weather
logic, next to allow them to transition (way after I get actual
gameplay)

I am now comfortable with the Title, I will move on to Nero, and the
Range, then I will come back around and make the Select, and its
transitions, from Title. and Splash, as well as make the Title to
Splash, Splash to Range, and Select to Range transitions.