The Range
Added the Environment to the Range.

The Environment 2

Clouds are complete, they have a falsified sub
scattering approximation
based on alpha, they look convincing enough to
warrant walking away from
them for now.

Time to move into the land of
getting the Range and Nero up and running.

Other exciting news!
The
cloud shader is perfect for Nero's paint to be done on meshes
in
realtime instead of on canvases, which is exciting.

Messed around
with some 4k clouds (they don't look sufficiently better
to warrant the
massive cost of them)

Moved some stuff around.