UNFIXED
So I undid all that stuff, and will do it again properly.
I ended up having the Sync project that I was working on, and another,
inside that one, and I wrongly concluded I had been working on the one
inside, and overwrote the original with that one, so here we are,
reverting that madness.

Now I will just rightly delete that one.