Lento
Lento is in the works, I have an idle, an attack, and a loop animated,
as well as a barrier model in place.
Lento has been reworked in the GDD as well.
My plan now that everything back end seems to be stable is to push a new
character every couple days, and se where we are in a few weeks
(capstone final) so ideally I want 5-10 characters done for then, and a
world (plaza or grid)