Title
So I decided to polish the title some more, I can much more easily see a
website coming from this design, it is superior. I may in fact remove
any version of the old title, and switch to a black text on white, with
colour splat accent, if I can.

I need to figure out how my UI is going to look in game, and I need to
begin making Lento do the do.